### 2021-08-31 #3662 @lucydot
Hi @jorisparet I'll be your editor for this submission 👋

Do you have any suggestions for potential reviewers? If so it'd be useful if you could mention them here in this thread (without tagging them with an @). In addition, [this list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
