### 2021-08-31 #3662 @jorisparet
Hi and thank you @lucydot for accepting to be editor for our paper! (and thank you @danielskatz for starting the thread so rapidly).

As reviewers, we would like to suggest **Fturci** and **Kappie** as they work in closely related fields and have great programming skills in Python.
