### 2021-09-01 #3662 @lucydot
Hi @govarguz and @fturci :wave: would you be able to review given your experience in soft matter and python?

@fturci I think this may be the first time you've been invited to review for JOSS (?) - I've sent you an email with a little more information
