### 2021-09-04 95c1929
Streamline a bit atooms and mdtraj parser
