### 2021-09-04 c82d40d
Add more tests on trajectories and helper functions
