### 2021-09-04 d5f5ba8
Fix test looking for networkx (typo in variable)
