### 2021-09-07 e431a79
Add CI and badges
