### 2021-09-12 ba9c497
Add flake8 configuration and fix some PEP8 issues
