+++ partycls/__init__.py
@@ -8,3 +8,4 @@ from .clustering import *
 from .dim_reduction import *
 from .feature_scaling import *
 from .helpers import *
+from.feature_selection import FeatureSelector
