+++ README.md
@@ -99,7 +99,16 @@ make install
 Support and contribution
 ------------------------
 
-If you wish to contribute or report an issue, feel free to [contact us](mailto:joris.paret@umontpellier.fr) or to use the [issue tracker](https://github.com/jorisparet/partycls/issues) and [pull requests](https://github.com/jorisparet/partycls/pulls) from the code repository.
+If you wish to contribute or report an issue, feel free to [contact us](mailto:joris.paret@umontpellier.fr) or to use the [issue tracker](https://github.com/jorisparet/partycls/issues) and [pull requests](https://github.com/jorisparet/partycls/pulls) from the [code repository](https://github.com/jorisparet/partycls).
+
+We largely follow the [GitHub flow](https://guides.github.com/introduction/flow/) to integrate community contributions. In essence:
+* fork the repository ;
+* create a feature branch from *master* ;
+* unleash your creativity ;
+* run the tests ;
+* open a pull request ;
+
+We also welcome contributions from other platforms, such as GitLab instances. Just let us know where to find your feature branch.
 
 Authors
 -------
