### 2021-09-13 20fb552
Add dynamic coverage badge (Coveralls)
