### 2021-09-13 #3662 @lucydot
Excellent ✨ -  I'll ask our bot Whedon to start a review issue thread (this is the _pre_-review thread). If you're new to the JOSS review process, be sure to read the instructions at the top of that new thread.

@govarguz @FTurci thank-you for agreeing to review. We're asking for reviews to be complete within a 4-6 week period. As the review is usually an iterative process (rather than something you do in one sitting) it's best for reviews to start earlier in that timeline.
