### 2021-09-13 #3723 @whedon
[REVIEW]: partycls: A Python package for structural clustering
**Submitting author:** @jorisparet (<a href="http://orcid.org/0000-0002-1300-1227">Joris Paret</a>)
**Repository:** <a href="https://github.com/jorisparet/partycls" target ="_blank">https://github.com/jorisparet/partycls</a>
**Version:** v1.1.1
