+++ paper/paper.bib
@@ -205,6 +205,7 @@ date = {2019-10-15}
 @article{Ganda_2020,
   title =	 {Concepts, fabrication methods and applications of living crystallization-driven self-assembly of block copolymers},
   journal =	 {Progress in Polymer Science},
+  DOI = 	 {10.1016/j.progpolymsci.2019.101195}
   author =	 {Ganda, Sylvia and Stenzel, Martina H.},
   year =	 2020,
   volume =	 101,
@@ -314,7 +315,7 @@ date = {2019-10-15}
     pages = {107275},
     year = {2020},
     issn = {0010-4655},
-    doi = {https://doi.org/10.1016/j.cpc.2020.107275},
+    doi = {10.1016/j.cpc.2020.107275},
     url = {http://www.sciencedirect.com/science/article/pii/S0010465520300916},
     keywords = {Simulation analysis, Molecular dynamics,
                 Monte Carlo, Computational materials science},
