### 2021-09-13 #3723 @lucydot
@FTurci @govarguz Thanks again for agreeing to review. Reviewer checklists have been generated for you above, feel free to get started whenever is convenient for you in the nearest few weeks. It is best to raise any issues during the review process in the [software repository](https://github.com/jorisparet/partycls/issues) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.

I'll check back in regularly, and do not hesitate to ask if you have questions.
