### 2021-09-13 #3723 @jorisparet
Hi @lucydot :wave:! The paper is still on the same branch *joss-paper*. Whedon had the same problem at the very beginning of the pre-review thread until the person who started the thread asked it to check this branch for references. I'll try that immediately.
