+++ paper/paper.bib
@@ -205,7 +205,7 @@ date = {2019-10-15}
 @article{Ganda_2020,
   title =	 {Concepts, fabrication methods and applications of living crystallization-driven self-assembly of block copolymers},
   journal =	 {Progress in Polymer Science},
-  DOI = 	 {10.1016/j.progpolymsci.2019.101195}
+  DOI = 	 {10.1016/j.progpolymsci.2019.101195},
   author =	 {Ganda, Sylvia and Stenzel, Martina H.},
   year =	 2020,
   volume =	 101,
