+++ partycls/feature_selection.py
@@ -1,58 +1,108 @@
 class FeatureSelector:
     
-    def __init__(self):
-        self.selections = []
-    
-    def show_feature_space(self, X_1, X_2, method='heatmap', show_selections=True, **kwargs):
+    def __init__(self, descriptor, X_1, X_2):
+        self.descriptor = descriptor
+        self.X_1 =  X_1
+        self.X_2 = X_2
+        self.active_selections = []
+        
+    def add_rectangular_selection(self, x0, y0, x1, y1, label=None, **kwargs):
+        '''
+        Rectangular selection.
+        
+        +----------(x1,y1)
+        |                |
+        |                |
+        |                |
+        (x0,y0)----------+
+        
+        '''
+        # Default name for label
+        if label is None:
+            label = 'S_{}'.format(len(self.selections))
+            
+        # Set particle property trajectory-wide to 0
+        if not hasattr(self.descriptor.trajectory[0].particle[0], label):
+            self.descriptor.trajectory.set_property(label, 0)
+        # Set property to 1 if particle is in the current selection
+        R = Rectangle(x0, y0, x1, y1, label=label, **kwargs)
+        count = 0
+        for frame in self.descriptor.groups[0]:
+            for particle in frame:
+                is_inside = int(R.is_inside(self.X_1[count], self.X_2[count]))
+                setattr(particle, label, is_inside)
+                count += 1
+            
+        # Add selection
+        self.active_selections.append(R)
+
+    def clear_selection(self, label):
+        '''
+        Clear all selections labeled as `label` and clear the associated
+        particle property from the trajectory.
+        '''
+        # Delete particle property trajectory-wide           
+        for system in self.descriptor.trajectory:
+            for particle in system.particle:
+                delattr(particle, label)
+        # Delete selection(s) from list of active selections
+        self.active_selections = [s for s in self.active_selections if s.label != label]
+
+    def clear_all_selections(self):
+        '''
+        Clear all selections their associated
+        particle properties from the trajectory.
+        '''
+        for selection in self.active_selections:
+            self.clear_selection(selection.label)
+
+    def show_feature_space(self, method='heatmap', show=True, show_selections=True, **kwargs):
+        '''
+        Show the feature space (X_1,X_2) in the form of a scatter plot or a
+        heatmap. Active selections can be superposed.
+        '''
         import matplotlib.pyplot as plt
         
+        fig = plt.figure()
         if method == 'heatmap':
-            plt.hist2d(X_1, X_2, **kwargs)
+            plt.hist2d(self.X_1, self.X_2, **kwargs)
         elif method == 'scatter':
-            plt.scatter(X_1, X_2, **kwargs)
+            plt.scatter(self.X_1, self.X_2, **kwargs)
         else:
             print('Unknown method.')
             
         if show_selections:
-            for S in self.selections:
+            for S in self.active_selections:
                 from matplotlib.patches import Rectangle as RectanglePatch
-                if S.shape == 'rectangle':
-                    shape = RectanglePatch((S.x, S.y), S.width, S.height, **S.kwargs)
+                if isinstance(S, Rectangle):
+                    anchor = (S.x0, S.y0)
+                    width = S.x1 - S.x0
+                    height = S.y1 - S.y0
+                    shape = RectanglePatch(anchor, width, height, **S.kwargs)
                 else:
                     print('Unknown shape.')
                 plt.gca().add_patch(shape)
-        plt.show()
+        if show:
+            plt.show()
+        return fig
         
-    def add_rectangular_selection(self, x, y, width, height, **kwargs):
-        R = Rectangle(x, y, width, height, **kwargs)
-        self.selections.append(R)
-
[…]
