### 2021-10-03 jorisparet/partycls#2 @FTurci
[minor] Clarify in the README which descriptors are available out of the box
The `Features` sections of the `README.md` should list the descriptors that are readily available within the package. At the moment, this information appears to be buried in tutorial 3.
