### 2021-10-03 jorisparet/partycls#3 @FTurci
[backend] Ovito support?
Running the snippet present in the README.md
```
from partycls import Trajectory, Workflow

traj = Trajectory('partycls/data/dislocation.xyz')
wf = Workflow(traj, descriptor='ba', clustering='kmeans')
wf.run()
traj[0].show(color='label', backend='ovito')
```

returns an error

```ValueError: unknown backend for visualization```
