### 2021-10-03 jorisparet/partycls#4 @FTurci
[testing] pytest issue with Fortran module?
Something odd with the tests: Running `pytest` in the `test` folder returns
```E   ModuleNotFoundError: No module named 'partycls.descriptor.realspace_wrap'```

Should I run a specific command for the test to include the Fortran module that `realspace_wrap` refers to? If so, this should be documented.
