### 2021-10-04 jorisparet/partycls#4 @FTurci
The only working route seems to be to pre-compile the fortran module with

```
cd partycls/descriptor
f2py -c -m realspace_wrap realspace.f90

```

and the execute the tests. I suggest to add this information to the documentation (fore example, the README).
