### 2021-10-04 jorisparet/partycls#3 @jorisparet
Hi Francesco,

I assume you installed the package via pypi? The support for Ovito was added here _after_ the first upload on pypi so the versions on GitHub and pypi slightly differ. I will increment the version to 1.1.0 and push it to pypi to account for the latest changes on the repo.

Thank you for raising this issue.
