### 2021-10-04 jorisparet/partycls#4 @jorisparet
Absolutely, the Fortran module should be compiled prior to running the tests. I will add a note to the README. Thanks!
