### 2021-10-04 jorisparet/partycls#2 @jorisparet
Good point, I will change that rapidly. Thank you!
