### 2021-10-04 fdb7614
List the native descriptors and add external links
