+++ README.md
@@ -67,7 +67,7 @@ Requirements
 * [numpy](https://pypi.org/project/numpy/)
 * [scikit-learn](https://scikit-learn.org)
 * [optional] [mdtraj](https://www.mdtraj.org) (additional trajectory formats)
-* [optional] [atooms](https://framagit.org/atooms/atooms) (additional trajectory formats)
+* [optional] [atooms](https://framagit.org/atooms/atooms) < 3.0.0 (additional trajectory formats)
 * [optional] [dscribe](https://singroup.github.io/dscribe) (additional descriptors)
 * [optional] [matplotlib](https://matplotlib.org/) (visualization)
 * [optional] [ovito](https://ovito.org/) (visualization)
