### 2021-10-04 a53b8b5
Add fortran target to Makefile
