### 2021-10-04 b9e5306
Add directives for running the tests
