### 2021-10-05 #3723 @FTurci
I  have gone through the checklist, raised a few issues and @jorisparet has already addressed them. I do not know if it is customary, but I would like to briefly summarise my assessment.

The repository is well prepared, the documentation is useful, the installation is easy. The article is well written and should help any future user. The package allows for more physical ingredients to be added through `dscribe` which is a very interesting expansion direction. The development of structure descriptors and their aggregation into meaningful physical objects to identify microscale features in dense materials is a developing area of research and this work fits nicely within this context.
