### 2021-10-05 #3723 @govarguz
> @FTurci has reported that there is a problem with editing the reviewer checklists at the top of this issue. @govarguz have you had a similar problem?
>
> I will try re-inviting @FTurci as reviewer and hopefully that will fix it - please let me know if it does not.

yes, I had the same problem could you please re-invite me?
