### 2021-10-06 bb8f881
Smoothed BA start
