### 2021-10-06 #3723 @lucydot
@FTurci that's great! Your summary is very welcome, thank you. I can see that you raised issues [here](https://github.com/jorisparet/partycls/issues/4), [here](https://github.com/jorisparet/partycls/issues/3) and [here](https://github.com/jorisparet/partycls/issues/2) and they are all closed (linking them from here so easier to trace back).
