+++ partycls/descriptor/__init__.py
@@ -4,6 +4,7 @@ Structural descriptors.
 
 from .descriptor import StructuralDescriptor, DummyDescriptor
 from .ba import BondAngleDescriptor
+from .smoothed_ba import SmoothedBondAngleDescriptor
 from .gr import RadialDescriptor
 from .bo import BondOrientationalDescriptor, LechnerDellagoDescriptor
 from .dscribe import *
