### 2021-10-14 c210e7a
Add note on dynamical clustering
