+++ paper/paper.bib
@@ -331,3 +331,15 @@ date = {2019-10-15}
 	eprinttype = {arxiv},
 	eprint = {1808.05292}
 }
+
+@collection{berthier_dynamical_2011,
+	title = {Dynamical Heterogeneities in Glasses, Colloids, and Granular Media},
+	isbn = {978-0-19-969147-0},
+	url = {https://oxford.universitypressscholarship.com/10.1093/acprof:oso/9780199691470.001.0001/acprof-9780199691470},
+	series = {International Series of Monographs on Physics},
+	pagetotal = {464},
+	publisher = {Oxford University Press},
+	editor = {Berthier, Ludovic and Biroli, Giulio and Bouchaud, Jean-Philippe and Cipelletti, Luca and van Saarloos, Wim},
+	date = {2011},
+	doi = {10.1093/acprof:oso/9780199691470.001.0001}
+}
