### 2021-10-14 6f962ee
Adapt docstring
