### 2021-10-14 6e4304f
Revise text to address #9 and #11
