+++ paper/paper.bib
@@ -343,3 +343,47 @@ date = {2019-10-15}
 	date = {2011},
 	doi = {10.1093/acprof:oso/9780199691470.001.0001}
 }
+
+@article{Guzman_Tretyakov_Kobayashi_Fogarty_Kreis_Krajniak_Junghans_Kremer_Stuehn_2019,
+  title =	 {ESPResSo++ 2.0: Advanced methods for multiscale
+                  molecular simulation},
+  volume =	 238,
+  ISSN =	 {0010-4655},
+  DOI =		 {10.1016/j.cpc.2018.12.017},
+  journal =	 {Computer Physics Communications},
+  author =	 {Guzman, Horacio V. and Tretyakov, Nikita and
+                  Kobayashi, Hideki and Fogarty, Aoife C. and Kreis,
+                  Karsten and Krajniak, Jakub and Junghans, Christoph
+                  and Kremer, Kurt and Stuehn, Torsten},
+  year =	 2019,
+  month =	 {May},
+  pages =	 {66–76}
+}
+
+@article{Weeks_Crocker_Levitt_Schofield_Weitz_2000,
+  title =	 {Three-Dimensional Direct Imaging of Structural
+                  Relaxation Near the Colloidal Glass Transition},
+  volume =	 287,
+  ISSN =	 {0036-8075, 1095-9203},
+  DOI =		 {10.1126/science.287.5453.627},
+  number =	 5453,
+  journal =	 {Science},
+  author =	 {Weeks, Eric R. and Crocker, J. C. and Levitt, Andrew
+                  C. and Schofield, Andrew and Weitz, D. A.},
+  year =	 2000,
+  month =	 {Jan},
+  pages =	 {627–631}
+}
+
+@article{Markidis_Peng_Podobas_Jongsuebchoke_Bengtsson_Herman_2020,
+  title =	 {Automatic Particle Trajectory Classification in
+                  Plasma Simulations},
+  url =		 {http://arxiv.org/abs/2010.05348},
+  note =	 {arXiv: 2010.05348},
+  journal =	 {arXiv:2010.05348 [physics]},
+  author =	 {Markidis, Stefano and Peng, Ivy and Podobas, Artur
+                  and Jongsuebchoke, Itthinat and Bengtsson, Gabriel
+                  and Herman, Pawel},
+  year =	 2020,
+  month =	 {Oct}
+}
