+++ paper/paper.bib
@@ -375,15 +375,13 @@ date = {2019-10-15}
   pages =	 {627–631}
 }
 
-@article{Markidis_Peng_Podobas_Jongsuebchoke_Bengtsson_Herman_2020,
-  title =	 {Automatic Particle Trajectory Classification in
-                  Plasma Simulations},
-  url =		 {http://arxiv.org/abs/2010.05348},
-  note =	 {arXiv: 2010.05348},
-  journal =	 {arXiv:2010.05348 [physics]},
-  author =	 {Markidis, Stefano and Peng, Ivy and Podobas, Artur
-                  and Jongsuebchoke, Itthinat and Bengtsson, Gabriel
-                  and Herman, Pawel},
-  year =	 2020,
-  month =	 {Oct}
+@INPROCEEDINGS{Markidis_Peng_Podobas_Jongsuebchoke_Bengtsson_Herman_2020,
+  author={Markidis, Stefano and Peng, Ivy and Podobas, Artur and Jongsuebchoke, Itthinat and Bengtsson, Gabriel and Herman, Pawel},
+  booktitle={2020 IEEE/ACM Workshop on Machine Learning in High Performance Computing Environments (MLHPC) and Workshop on Artificial Intelligence and Machine Learning for Scientific Applications (AI4S)}, 
+  title={Automatic Particle Trajectory Classification in Plasma Simulations}, 
+  year={2020},
+  volume={},
+  number={},
+  pages={64-71},
+  doi={10.1109/MLHPCAI4S51975.2020.00014}
 }
