+++ paper/paper.bib
@@ -385,3 +385,15 @@ date = {2019-10-15}
   pages={64-71},
   doi={10.1109/MLHPCAI4S51975.2020.00014}
 }
+
+@article{anderson_hoomd_2020,
+title = {HOOMD-blue: A Python package for high-performance molecular dynamics and hard particle Monte Carlo simulations},
+journal = {Computational Materials Science},
+volume = {173},
+pages = {109363},
+year = {2020},
+issn = {0927-0256},
+doi = {https://doi.org/10.1016/j.commatsci.2019.109363},
+url = {https://www.sciencedirect.com/science/article/pii/S0927025619306627},
+author = {Joshua A. Anderson and Jens Glaser and Sharon C. Glotzer}
+}
