### 2021-10-20 9fbf431
Remove reference on confocal microscopy
