+++ paper/paper.bib
@@ -360,21 +360,6 @@ date = {2019-10-15}
   pages =	 {66–76}
 }
 
-@article{Weeks_Crocker_Levitt_Schofield_Weitz_2000,
-  title =	 {Three-Dimensional Direct Imaging of Structural
-                  Relaxation Near the Colloidal Glass Transition},
-  volume =	 287,
-  ISSN =	 {0036-8075, 1095-9203},
-  DOI =		 {10.1126/science.287.5453.627},
-  number =	 5453,
-  journal =	 {Science},
-  author =	 {Weeks, Eric R. and Crocker, J. C. and Levitt, Andrew
-                  C. and Schofield, Andrew and Weitz, D. A.},
-  year =	 2000,
-  month =	 {Jan},
-  pages =	 {627–631}
-}
-
 @INPROCEEDINGS{Markidis_Peng_Podobas_Jongsuebchoke_Bengtsson_Herman_2020,
   author={Markidis, Stefano and Peng, Ivy and Podobas, Artur and Jongsuebchoke, Itthinat and Bengtsson, Gabriel and Herman, Pawel},
   booktitle={2020 IEEE/ACM Workshop on Machine Learning in High Performance Computing Environments (MLHPC) and Workshop on Artificial Intelligence and Machine Learning for Scientific Applications (AI4S)}, 
