+++ paper/paper.md
@@ -40,7 +40,7 @@ By grouping the particles according to the similarity of their local structure, 
 With the present code, we aim to provide a coherent numerical framework for unsupervised learning of structural and dynamical features of condensed matter systems.
 To the best of our knowledge, there is currently no publicly available code that provides and integrates all the necessary tools needed for this kind of analysis.
 Through a variety of structural descriptors, dimensionality reduction methods, clustering algorithms and filtering options, partycls makes it possible to discover the key structural features of a system and to assess the robustness of the results.
-It can readily analyze trajectories generated by molecular dynamics simulation softwares, both within particle-based and multi-scale approaches (see *e.g.* @plimpton_fast_1995, @anderson_hoomd_2020, @Guzman_Tretyakov_Kobayashi_Fogarty_Kreis_Krajniak_Junghans_Kremer_Stuehn_2019), but also experimental data obtained with confocal microscope techniques [@Weeks_Crocker_Levitt_Schofield_Weitz_2000].
+It can readily analyze trajectories generated by molecular dynamics simulation softwares, both within particle-based and multi-scale approaches (see *e.g.* @plimpton_fast_1995, @anderson_hoomd_2020 and @Guzman_Tretyakov_Kobayashi_Fogarty_Kreis_Krajniak_Junghans_Kremer_Stuehn_2019).
 The code has already been used in the context of a recent publication [@paret_assessing_2020] and can be easily extended.
 In particular, future versions will implement descriptors that depend on both space *and* time, to learn about the dynamics of the system as well.
 Spatio-temporal clustering may provide valuable insight into the heterogeneous dynamics of amorphous materials [@berthier_dynamical_2011] and into particles' trajectories in more extreme states of matter [@Markidis_Peng_Podobas_Jongsuebchoke_Bengtsson_Herman_2020].
