+++ paper/paper.bib
@@ -378,7 +378,7 @@ volume = {173},
 pages = {109363},
 year = {2020},
 issn = {0927-0256},
-doi = {https://doi.org/10.1016/j.commatsci.2019.109363},
+doi = {10.1016/j.commatsci.2019.109363},
 url = {https://www.sciencedirect.com/science/article/pii/S0927025619306627},
 author = {Joshua A. Anderson and Jens Glaser and Sharon C. Glotzer}
 }
