+++ partycls/descriptor/__init__.py
@@ -7,4 +7,5 @@ from .ba import BondAngleDescriptor
 from .smoothed_ba import SmoothedBondAngleDescriptor
 from .gr import RadialDescriptor
 from .bo import BondOrientationalDescriptor, LechnerDellagoDescriptor
+from .smoothed_bo import SmoothedBondOrientationalDescriptor
 from .dscribe import *
