+++ setup.py
@@ -18,7 +18,7 @@ args = dict(name='partycls',
             long_description=readme,
             long_description_content_type="text/markdown",
             author='Joris Paret',
-            author_email='joris.paret@umontpellier.fr',
+            author_email='joris.paret@gmail.com',
             packages=['partycls',
                       'partycls/core',
                       'partycls/descriptor'],
@@ -30,7 +30,10 @@ args = dict(name='partycls',
                 'Development Status :: 5 - Production/Stable',
                 'Intended Audience :: Science/Research',
                 'Programming Language :: Python :: 3',
+                'Programming Language :: Python :: 3.6',                
                 'Programming Language :: Python :: 3.7',
+                'Programming Language :: Python :: 3.8',
+                'Programming Language :: Python :: 3.9',
                 'Topic :: Scientific/Engineering :: Physics'
                 ]
 )
