### 2021-11-03 #3723 @jorisparet
Thank you very much, @lucydot! And thank you again for your reviews, @FTurci and @govarguz.

As requested, I just pushed a [new tagged version](https://github.com/jorisparet/partycls/tree/1.1.1) (1.1.1) and archived it to [zenodo](https://doi.org/10.5281/zenodo.5639407). Here is the corresponding DOI:

**10.5281/zenodo.1183302**

Let us know if we can do anything else. :slightly_smiling_face:
