### 2021-11-03 4375d21
Add jupyter book support for tutorial
