+++ README.md
@@ -132,4 +132,4 @@ Authors
 
 Joris Paret
 
-Daniele Coslovich: http://www-dft.ts.infn.it/~coslovich/
+Daniele Coslovich: https://www.units.it/daniele.coslovich/
