### 2021-11-08 #3723 @lucydot
Thanks @jorisparet though I think the DOI is: 10.5281/zenodo.5639407 (a mixup with another package of yours I think)
